+++ src/lv3/OperatorType.java
@@ -2,22 +2,22 @@ package lv3;
 
 public enum OperatorType {
     ADD("+") {
-        public int cal(int a, int b) {
+        public double cal(double a, double b) {
             return a + b;
         }
     },
     SUBTRACT("-") {
-        public int cal(int a, int b) {
+        public double cal(double a, double b) {
             return a - b;
         }
     },
     MULTIPLY("*") {
-        public int cal(int a, int b) {
+        public double cal(double a, double b) {
             return a * b;
         }
     },
     DIVIDE("/") {
-        public int cal(int a, int b) {
+        public double cal(double a, double b) {
             if (b == 0) {
                 System.out.println("나눗셈 연산에서 분모(두 번째 정수)에 0이 입력될 수 없습니다.");
                 return 0; // 또는 return -1 등으로 예외 상황임을 알리는 값 반환
@@ -38,7 +38,7 @@ public enum OperatorType {
         return symbol;
     }
 
-    public abstract int cal(int a, int b);
+    public abstract double cal(double a, double b);
     public static OperatorType fromSymbol(String symbol) {
         for (OperatorType op : OperatorType.values()) {
             if (op.getSymbol().equals(symbol)) {
